533-e4-django-search/533e5-dkrs-search/djangosite/requirements.txt
#Django>=2,<3
Django<=3.2
django-extensions
djangorestframework
djangorestframework_simplejwt
django-cors-headers
#django-crontab
apscheduler
gunicorn